requirements.txt
Flask==3.1.0
flask-cors==5.0.1
opencv-python==4.11.0.86
numpy==2.0.2
deepface==0.0.93
tensorflow==2.18.0
keras==3.8.0
pandas==2.2.3
matplotlib
gdown==5.2.0
requests==2.32.3
retina-face==0.0.17
mtcnn==0.1.1
gunicorn==23.0.0
tqdm==4.67.1
fire==0.7.0
termcolor==2.5.0
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
importlib-metadata==8.6.1
Jinja2==3.1.5
Werkzeug==3.1.3
filelock==3.17.0
beautifulsoup4==4.13.3
packaging==24.2
optree==0.14.1
rich==13.9.4
h5py==3.13.0
ml-dtypes==0.5.1
absl-py==2.1.0
namex==0.0.8
pytz==2025.1
python-dateutil==2.9.0.post0
tzdata==2025.1
idna==3.10
charset-normalizer==3.4.1
urllib3==2.3.0
certifi==2025.1.31
mysql-connector-python==9.2.0
mediapipe